New additions:
- error handling - I want to have a separate tab that only users with admin rights can view which displays any scripts that
are running with errors

- I want to add a tab on the side bar titled ‘tape summary ‘ which once selected can display a summary of various models
already running on the site. The idea is that if I wanted to pick say 5 scripts - get the results from those scripts and displays them all on one place. If you look at the script ‘tape summary ‘ it begins to try and complete this task. I would like the model summary to be able to display both a graph and a table on the same tab

- on the main script page I want to be able to click through and sort scripts by last created

- be able to drill down through the category tree in order to quickly sort scripts—- so maybe a user clicks tape - breadth - and they can quickly see all the different scripts.